+++ paper/paper.bib
@@ -3,7 +3,7 @@
   author={Vicsek, T and Czirók, A and Ben-Jacob, E and Cohen, I and Shochet, O},
   journal={Phys Rev Lett},
   year={2016},
-  doi={https://doi.org/10.1103/PhysRevLett.75.1226},
+  doi={10.1103/PhysRevLett.75.1226},
 }
 
 @article{reif:1999,
@@ -11,7 +11,7 @@
   author={Reif, JH and Wang, H},
   journal={Robot. Auton. Syst.},
   year={1999},
-  doi={https://doi.org/10.1016/S0921-8890(99)00004-4},
+  doi={10.1016/S0921-8890(99)00004-4},
 }
 
 @article{spears:1999,
@@ -19,7 +19,7 @@
   author={Spears, WM and Gordon, DF},
   journal={International Conference on Information Intelligence and Systems},
   year={1999},
-  doi={https://doi.org/10.1109/ICIIS.1999.810278},
+  doi={10.1109/ICIIS.1999.810278},
 }
 
 @article{helbing:2000,
@@ -27,7 +27,7 @@
   author={Helbing, D, Farkas, I and Vicsek, T},
   journal={Nature},
   year={2000},
-  doi={https://doi.org/10.1038/35035023},
+  doi={10.1038/35035023},
 }
 
 @article{leonard:2001,
@@ -35,7 +35,7 @@
   author={Leonard, NE and Fiorelli, E},
   journal={IEEE Conference on Decision and Control},
   year={2001},
-  doi={https://doi.org/10.1109/CDC.2001.980728},
+  doi={10.1109/CDC.2001.980728},
 }
 
 @article{pinciroli:2008,
@@ -43,7 +43,7 @@
   author={Pinciroli, Carlo and Birattari, Mauro and Tuci, Elio and Dorigo, Marco and del Rey Zapatero, Marco and Vinko, Tamas and Izzo, Dario},
   journal={Ant Colony Optimization and Swarm Intelligence},
   year={2008},
-  doi={https://doi.org/10.1007/978-3-540-87527-7_36},
+  doi={10.1007/978-3-540-87527-7_36},
 }
 
 @article{jadbabaie:2003,
@@ -51,7 +51,7 @@
   author={Jadbabaie, A and Jie, Lin and Morse, AS},
   journal={IEEE Trans Automat Contr},
   year={2003},
-  doi={https://doi.org/10.1109/TAC.2003.812781},
+  doi={10.1109/TAC.2003.812781},
 }
 
 @article{couzin:2005,
@@ -59,7 +59,7 @@
   author={Couzin, ID and Krause, J and Franks, NR and Levin, SA},
   journal={Nature},
   year={2005},
-  doi={https://doi.org/10.1038/nature03236},
+  doi={10.1038/nature03236},
 }
 
 @article{freeman:2009,
@@ -67,7 +67,7 @@
   author={Freeman, R and Biro, D},
   journal={J Navigation},
   year={2009},
-  doi={https://doi.org/10.1017/S0373463308005080},
+  doi={10.1017/S0373463308005080},
 }
 
 @article{chamanbaz:2017,
@@ -75,7 +75,7 @@
   author={Chamanbaz, M and Mateo, D and Zoss, BM and Tokić, G and Wilhelm, E and Bouffanais, R and et al.},
   journal={Front Robot AI},
   year={2017},
-  doi={https://doi.org/10.3389/frobt.2017.00012},
+  doi={10.3389/frobt.2017.00012},
 }
 
 @article{zoss:2018,
@@ -83,7 +83,7 @@
[…]
