### 2023-05-12 #5462 @mrsonandrade
> @mrsonandrade, could you take a look at those DOI error messages above?

Sure! I've fixed all of them.
