### 2023-07-10 #5647 @editorialbot
[REVIEW]: PySwarming: a research toolkit for Swarm Robotics
**Submitting author:** @mrsonandrade (<a href="http://orcid.org/0000-0002-5023-8733">Emerson Andrade</a>)
**Repository:** https://github.com/mrsonandrade/pyswarming
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.3
