### 2023-07-15 mrsonandrade/pyswarming#1 @sea-bass
Distribution around deployment point is not configurable
I was confused by the initial example that setting `deployment_point=[0.0, 0.0]` automatically adds a uniform distribution of +/- 5.0 meters around those coordinates, plus an orientation in the full range of 0 to 2*pi radians.

Looking at the source code, I see that `Swarm._create_robots()` simply hard-codes these parameters.

As such, I would suggest the following:
* [MUST HAVE] Allow users to configure the bounds of the X/Y position and angle
* [NICE TO HAVE] Allow users to switch between uniform distribution, Gaussian distribution, no distribution, or a list of starting poses

https://github.com/openjournals/joss-reviews/issues/5647
