### 2023-07-15 mrsonandrade/pyswarming#4 @sea-bass
[JOSS Review] Paper Feedback
Thank you for submitting pyswarming to JOSS! As one of your reviewers, I'll collect a few issues here with the paper submission itself.

### General Comments
I think the "Related Software Packages" section should come in earlier in the paper, before you introduce PySwarming. This should serve as part of your introduction leading to the "meat" of the paper, and meet the JOSS expectation below:

* **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

Do the authors have any other published work that uses PySwarming? If so, it's a great opportunity to include this in the references, if it has not yet been added.

Lastly, the paper is missing a conclusion! Please add one.

---

### Summary
The summary seems like it abruptly switches from an introduction to swarm robotics to a list of things that PySwarming can do. Consider making this a bit more gentle by following a layout like this:

* Introduce swarm robotics
* Identify the need for standard tooling for development/benchmarking/etc.
* Explicitly say something like "we introduce PySwarming, a tool that does XYZ"
* Briefly say what the tool does (see next paragraph) and what this paper covers.

On the note of "briefly say what the tool does", I think in your summary the citations to leaderless coordination, preferred direction, aggregation, etc. Consider talking about these algorithms in a separate section. Most papers will have

* Abstract: one paragraph similar to your summary, but a bit more brief and with little to no citations
* Introduction: A few paragraphs that perform a brief literature review (where these citations could go), plus the current state of the art methods. In the context of JOSS, this is where you could talk about related software packages such as Buzz.

---

### Example Usage

I like this section! The only thing I would suggest here is maybe adding more details about how a user might use this tool to create their own new behaviors. The paper advertises that PySwarming is flexible and easy to extend, so it's worthwhile to talk about that a little bit.

---

### Related Software Packages

* Typo: the first paragraph refers to "PySwarm", change to "PySwarming".
* I think this list could probably be expanded a little bit. I'm sure there are more than just the few resources pointed out, see e.g. this paper: https://www.mdpi.com/2079-3197/10/6/80
* If you have mentioned Carlo Pinciroli's Buzz programming language, you should also mention his ARGoS simulator: https://www.argos-sim.info/
* You should not mention your own contribution in this section, so you can remove that 3rd paragraph and move it to sections where you are talking about PySwarming.

https://github.com/openjournals/joss-reviews/issues/5647
