### 2023-07-15 #5647 @sea-bass
I have gone through and reviewed this submission.

Overall, I think this is a really nice tool and a welcome addition to the swarm robotics community. As someone who has created simular multi-robot simulators (https://github.com/mathworks-robotics/mobile-robotics-simulation-toolbox and https://github.com/sea-bass/pyrobosim), I particularly appreciate that PySwarming isn't just a generic point robot simulator like these tools, but actually is specifically focused to implementing swarm behaviors from the literature in a modular framework.

PySwarming was extremely easy to install (I appreciate that it is `pip` installable). It was also easy to dig into the Python code and make changes/modifications. Finally, the API documentation and connection to ReadTheDocs was solid.

The code examples were helpful, and I really appreciated that the docs page has a nice Jupyter notebook that renders and shows the animations on the ReadTheDocs page. However, that same notebook didn't run standalone and the main `README.md` doesn't have self-contained examples beyond the first few simple ones. As such, I highly recommend also including smaller, self-contained examples in plain text `.py` format in the examples directory. My detailed feedback is here: https://github.com/mrsonandrade/pyswarming/issues/2

The paper is overall in good shape, but I think there are a few structural changes needed. Additionally, some more information on the state of swarm robotics software and on the developer workflows to add their own swarm behaviors to the tool would be helpful. Finally, a conclusion is needed. My detailed feedback is here: https://github.com/mrsonandrade/pyswarming/issues/4

In addition, I found a few more blocking software issues that I think should be resolved:
* The mean of the initial swarm positions is configurable, but not the distribution parameters themselves: https://github.com/mrsonandrade/pyswarming/issues/1
* The unit tests exist, but they are mostly checking that "something happened" rather than whether it happened correctly: https://github.com/mrsonandrade/pyswarming/issues/3

Thank you for submitting this exciting tool, and I look forward to seeing these comments addressed.
