### 2023-07-21 77831f8
Minor updates
The Swarm class is more configurable now; the documentation was made more clear; global variables were removed from the examples; addition of examples in plain text (.py); addition of unit new tests.
