+++ paper/paper.bib
@@ -1,17 +1,119 @@
 @article{vicsek:1995,
-  title={Novel Type of Phase Transition in a System of Self-Driven Particles},
-  author={Vicsek, T and Czirók, A and Ben-Jacob, E and Cohen, I and Shochet, O},
-  journal={Phys Rev Lett},
-  year={2016},
-  doi={10.1103/PhysRevLett.75.1226},
+  title={Novel type of phase transition in a system of self-driven particles},
+  author={Vicsek, Tam{\'a}s and Czir{\'o}k, Andr{\'a}s and Ben-Jacob, Eshel and Cohen, Inon and Shochet, Ofer},
+  journal={Physical review letters},
+  volume={75},
+  number={6},
+  pages={1226},
+  year={1995},
+  publisher={APS},
+  doi={10.1103/PhysRevLett.75.1226}
+}
+
+@article{vaughan:2008,
+  title={Massively multi-robot simulation in stage},
+  author={Vaughan, Richard},
+  journal={Swarm intelligence},
+  volume={2},
+  pages={189--208},
+  year={2008},
+  publisher={Springer},
+  doi={10.1007/s11721-008-0014-4}
+}
+
+@article{mondada:2004,
+  title={SWARM-BOT: A new distributed robotic concept},
+  author={Mondada, Francesco and Pettinaro, Giovanni C and Guignard, Andre and Kwee, Ivo W and Floreano, Dario and Deneubourg, Jean-Louis and Nolfi, Stefano and Gambardella, Luca Maria and Dorigo, Marco},
+  journal={Autonomous robots},
+  volume={17},
+  pages={193--221},
+  year={2004},
+  publisher={Springer},
+  doi={10.1023/B:AURO.0000033972.50769.1c}
+}
+
+@inproceedings{carpin:2007,
+  title={USARSim: a robot simulator for research and education},
+  author={Carpin, Stefano and Lewis, Mike and Wang, Jijun and Balakirsky, Stephen and Scrapper, Chris},
+  booktitle={Proceedings 2007 IEEE International Conference on Robotics and Automation},
+  pages={1400--1405},
+  year={2007},
+  organization={IEEE},
+  doi={10.1109/ROBOT.2007.363180}
+}
+
+@book{balch:1998teambots,
+  title={Behavioral diversity in learning robot teams},
+  author={Balch, Tucker R},
+  year={1998},
+  publisher={Georgia Institute of Technology}
+}
+
+@inproceedings{pinciroli:2011,
+  title={ARGoS: a modular, multi-engine simulator for heterogeneous swarm robotics},
+  author={Pinciroli, Carlo and Trianni, Vito and O'Grady, Rehan and Pini, Giovanni and Brutschy, Arne and Brambilla, Manuele and Mathews, Nithin and Ferrante, Eliseo and Di Caro, Gianni and Ducatelle, Frederick and others},
+  booktitle={2011 IEEE/RSJ International Conference on Intelligent Robots and Systems},
+  pages={5027--5034},
+  year={2011},
+  organization={IEEE},
+  doi={10.1109/IROS.2011.6094829}
+}
+
+@article{calderon:2022,
+  title={Swarm robotics: Simulators, platforms and applications review},
+  author={Calder{\'o}n-Arce, Cindy and Brenes-Torres, Juan Carlos and Solis-Ortega, Rebeca},
+  journal={Computation},
+  volume={10},
+  number={6},
+  pages={80},
+  year={2022},
+  publisher={MDPI},
[…]
