### 2023-07-21 mrsonandrade/pyswarming#1 @mrsonandrade
Great suggestions!
I have done both:
- Allow users to configure the bounds of the X/Y position and angle;
- Allow users to switch between uniform distribution, Gaussian distribution, no distribution, or a list of starting poses.
