### 2023-07-21 mrsonandrade/pyswarming#2 @mrsonandrade
Thank you for your feedback!
I have worked on all the points:
- Document the general approach a little better and give users some direction on when to use which approach;
- You are already using named arguments, and I could argue that this could be made clearer;
- No globals;
- Avoid single-letter variables (I have noticed that I forgot to avoid that on the README, I'll do it now);
- Provide More Variety of Examples;
- Small issues.
