### 2023-07-21 mrsonandrade/pyswarming#3 @mrsonandrade
Great!
I have implemented the point "verify that the two robots moved to the expected new poses" in the unit tests.
