### 2023-07-21 mrsonandrade/pyswarming#4 @mrsonandrade
@sea-bass, thank you very much for all the feedback.
I have worked on all the points that you have mentioned above:
- General Comments;
- Summary;
- Example Usage;
- Related Software Packages;

Let me know what you think.
And again, thank you for your time!
