+++ docs/source/conf.py
@@ -52,6 +52,7 @@ templates_path = ["_templates"]
 # This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = []
 
+nbsphinx_allow_errors = True
 
 # -- Options for HTML output -------------------------------------------------
 
