+++ pyswarming/__init__.py
@@ -34,6 +34,6 @@ __all__ = behaviors.__all__.copy()
 __all__ = swarm.__all__.copy()
 
 name = "pyswarming"
-__version__ = "1.1.0"
+__version__ = "1.1.1"
 __author__ = "Emerson Martins de Andrade, Antonio Carlos Fernandes and Joel Sena Sales Jr"
 __author_email__ = "mrson@oceanica.ufrj.br"
