+++ docs/source/conf.py
@@ -29,7 +29,7 @@ copyright = "2023, pyswarming developers"
 author = "pyswarming developers"
 
 # The full version, including alpha/beta/rc tags
-release = "1.1.0"
+release = "1.1.1"
 
 
 # -- General configuration ---------------------------------------------------
