+++ README.md
@@ -2,7 +2,7 @@
 [![Tests](https://github.com/mrsonandrade/pyswarming/actions/workflows/tests_package.yml/badge.svg)](https://github.com/mrsonandrade/pyswarming/actions/workflows/tests_package.yml)
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![Documentation Status](https://readthedocs.org/projects/pyswarming/badge/?version=latest)](https://pyswarming.readthedocs.io/en/latest/?badge=latest)
-![version](https://img.shields.io/badge/version-1.1.0-blue)
+![version](https://img.shields.io/badge/version-1.1.1-blue)
 [![Downloads](https://static.pepy.tech/badge/pyswarming)](https://pepy.tech/project/pyswarming)
 
 <img align="left" src="docs/readme_pics/logo.png">
@@ -80,42 +80,51 @@ This library includes the following algorithms to be used in swarm robotics:
 ## Examples using pyswarming.swarm
 ```python
 # importing the swarm creator
-import pyswarming.swarm as sw
+import pyswarming.swarm as ps
 ```
 
 ### Repulsion 
 ```python
-my_swarm = sw.Swarm(10, # number of robots
-                    0.5, # linear speed of each robot
-                    1.0, # sampling time
-                    [0.0, 0.0], # robots deployed randomly around x = 0.0, y = 0.0 (+- 5.0 meters)
-                    [[-50.0, 50.0], [-50.0, 50.0]], # plot limits x_lim, y_lim
-                    ['repulsion']) # list of behaviors
+# creating the swarm
+my_swarm = ps.Swarm(n = 10, # number of robots
+                    linear_speed = 0.5, # linear speed of each robot
+                    dT = 1.0, # sampling time
+                    deployment_point_limits = [[0.0, 0.0, 0.0], [5.0, 5.0, 0.0]], # lower and upper limits for the position deployment
+                    deployment_orientation_limits = [[0.0, 0.0, 0.0], [0.0, 0.0, 2*3.1415]], # lower and upper limits for the orientation deployment
+                    distribution_type =  'uniform', # type of distribution used to deploy the robots
+                    plot_limits = [[-50.0, 50.0], [-50.0, 50.0]], # plot limits x_lim, y_lim
+                    behaviors = ['repulsion']) # list of behaviors
 my_swarm.simulate()
 ```
 
 ### Collective navigation 
 ```python
-my_swarm = sw.Swarm(10, # number of robots
-                    0.5, # linear speed of each robot
-                    1.0, # sampling time
-                    [0.0, 0.0], # robots deployed randomly around x = 0.0, y = 0.0 (+- 5.0 meters)
-                    [[-50.0, 50.0], [-50.0, 50.0]], # plot limits x_lim, y_lim
-                    ['repulsion']) # list of behaviors
+# creating the swarm
+my_swarm = ps.Swarm(n = 10, # number of robots
+                    linear_speed = 0.5, # linear speed of each robot
+                    dT = 1.0, # sampling time
+                    deployment_point_limits = [[0.0, 0.0, 0.0], [5.0, 5.0, 0.0]], # lower and upper limits for the position deployment
+                    deployment_orientation_limits = [[0.0, 0.0, 0.0], [0.0, 0.0, 2*3.1415]], # lower and upper limits for the orientation deployment
[…]
