### 2023-07-22 mrsonandrade/pyswarming#2 @sea-bass
Looks great -- thanks for doing this!

On the readthedocs page, when you are using `simulate()` there is no animation that appears. But I'm fine with that if there is a limitation in the Jupyter notebooks render.
