### 2023-07-22 mrsonandrade/pyswarming#4 @sea-bass
Somehow the editorialbot on the original issue isn't updating the paper, but I took a look at the updated markdown file for now.

It's looking good, I still have some additional comments.

## Introduction
* I think generally, the introduction could do a better job of differentiating between the types of software related to swarm robotics. Particularly in "Related Software Packages" section, there is room to organize the software between simulators (ARGoS, Stage, etc.) and packages for behaviors (Buzz, ROS2Swarm, etc.). This will also let you more precisely position PySwarming in this category.
* You need to mention what need PySwarming meets. It's not the entire set of hardware+software needs, so what is it? My impression is that Python is arguably one of the most accessible languages for learning and prototyping today, so having a toolbox of standard swarm behavior primitives that someone can get going in < 1 hour can help lower the barrier for newcomers to the field, plus help more expert practitioners do basic prototyping of behaviors before trying to deploy them to embedded hardware. I think you have a more vague version of this in the "PySwarming" section, but it might be better served here instead.
* I also think you can explicitly say that a lot of the tools above, even if they may be capable, are NOT written in Python, so that in itself is a need being met by this tool!
* The statement of need is pretty sparse in general. Consider rolling in some of the discussions in the previous bullet point into that paragraph?
* Lastly, the Introduction section should also have some sentence saying "because of ABC, we introduce PySwarming, which does XYZ"

## Example Usage
* Instead of "mid and final results", say "intermediate and final results".
* I also think this example should have a repulsion component so that the "final" image doesn't show all the robots stacked on top of each other. It looks better visually, but also makes your example more compelling by adding a 3rd behavior primitive!
* My original comment on this section still persists: Suggest adding more details about how a user might use this tool to create their own new behaviors. The paper advertises that PySwarming is flexible and easy to extend, so it's worthwhile to talk about that a little bit. You link to examples, but just a few sentences describing the approach of adding new behaviors would go a long way.
