### 2023-07-22 #5647 @JHartzer
I have a concern regarding substantial scholarly effort of this work. As it currently stands, most of the lines of python code are tests, setups, or examples. While all these things are good to have, what remains are about 400 unique lines of code. Additionally, a large portion of this code is `auto_differentiation.py` which is used to compute function gradients. Excluding this leaves ~270 lines of code within `swarm.py` and `behaviors.py`, which in my opinion does not constitute a substantial scholarly effort.

I can understand that, by nature, swarm robotics has a large variety of objective functions to perform so creating an all-encompassing work is untenable. Other works have focused on specific algorithms such as [search and path planning](https://github.com/xzlxiao/SwarmRobotics), or are generic but contain high-level examples, like [ARGoS](http://argos-sim.info/index.php). Because this work currently only implements low-level behaviors, I think this currently falls short of the statement of need.
