### 2023-07-23 #5647 @mrsonandrade
Dear @JHartzer, I kind of understand your points. However, all these behaviors took me months of work, in which I had to read, understand, and select papers, and put everything in the same nomenclature and format in order to have a common ground. After that, I had to implement while trying to maintain the clearness and connection with their original sources (the articles), all in order to facilitate the understanding for those that are initiating/interested. Also, based on previous experiences, I tried to have good documentation to make it accessible to new researchers. So, please, take this into your scholarly effort considerations.

Regarding the “behaviors”, is exactly what you said, we have “a large variety of objective functions to perform”, but the main point of this library is joining the maximum behaviors/algorithms in order to avoid “repeated” use of those same behaviors/algorithms as something new. For instance, you probably already saw some work that uses a “behavior X” inspired by something, but you knew that the behavior was not new, it was just the same proposed by someone else with a different mathematical expression. So, the main point here is always the algorithms' standardization and implementation in PySwarming (and the use of Python language which needs no words). Also, you mentioned [search and path planning](https://github.com/xzlxiao/SwarmRobotics), but they have a Python simulator like [marabunta](https://github.com/david-mateo/marabunta), so, PySwarming here is not just another simulator, it comes as a package with various behaviors/algorithms that can be used even within these aforementioned packages. So, I hope you have got my point and I apologize if those points were not clear through the paper.
