### 2023-07-24 mrsonandrade/pyswarming#2 @mrsonandrade
Good! To manage that I have included an argument in `simulate()`, so, now they are able to appear on the readthedocs page 😃
