+++ paper/paper.md
@@ -29,29 +29,80 @@ bibliography: paper.bib
 
 # Summary
 
-When considering a system composed by a group of robots, swarm robotics is an approach that can be used to coordinating this group. These swarm robotic systems can be inspired or not by social insects or other animal societies [@trianni:2008], where basic behaviors are usually used to compose complex tasks. These previous mentioned basic behaviors have been studied for a long time, with applications, for example, to flocks, herds, and schools [@reynolds:1987] and multi-robot teams [@balch:1998].
+When considering a system composed of a group of robots, swarm robotics is an approach that can be used to coordinate this group. These swarms can be inspired or not by social insects or other animal societies [@trianni:2008], where basic behaviors are usually used to compose complex tasks. These previously mentioned basic behaviors have been studied for a long time, with applications, for example, to flocks, herds, and schools [@reynolds:1987] and multi-robot teams [@balch:1998].
 However, software standardization is urgently needed to maximize the applicability of swarm robotics to real-world applications [@nedjah:2019]. Here we introduce PySwarming, a tool that makes easy the coordination of swarms and serves as a centerpiece, organizing different methods developed in the swarm robotics field. Its flexibility (written in Python) and customizability (easily customized by users) encourage interaction and scientific progress in the research community.
 
 # Introduction
 
-Controlling a system composed of a group of robots can be a challenge, then, swarm robotics is an approach that is widely used to coordinate this kind of system. Furthermore, the changes that the field of swarm robotics has experienced in the last decade are unprecedented, with various demonstrations showing the potential of this technology [@drigo:2021]. Moreover, it is important to note that swarm robotics is a subfield of multi-robot systems, itself a subfield of mobile robot research [@dias:2021].
-Additionally, these swarming robotic systems may or may not be inspired by social insects and other animal societies [@trianni:2008], where basic behaviors are typically used to compose complex tasks. The aforementioned behavior has long been studied and applied, for instance, to herds, flocks, schools, etc. ([@reynolds:1987], [@toner:1998]), self-driven particles [@vicsek:1995], large collections of robots ([@reif:1999], [@spears:1999]), and the multi-robot teams [@balch:1998].
-
-## Statement of Need
-
-Although, as stated by [@nedjah:2019], a significant issue in the swarm robotics field is the lack of common generic hardware platforms and controller design methodologies. Also, the authors claim an urgent need for standardization of software, to allow a possible flourishing of swarm robotics applicability to real-world applications.
+Controlling a system composed of a group of robots can be a challenge, then, swarm robotics is an approach that is widely used to coordinate this kind of system. Furthermore, the changes that the field of swarm robotics has experienced in the last decade are unprecedented, with various demonstrations showing the potential of this technology [@dorigo:2021]. Moreover, it is important to note that swarm robotics is a subfield of multi-robot systems, itself a subfield of mobile robot research [@dias:2021].
+Additionally, these swarms may or may not be inspired by social insects and other animal societies [@trianni:2008], where basic behaviors are typically used to compose complex tasks. The aforementioned behaviors have long been studied and applied, for instance, to herds, flocks, schools, etc. ([@reynolds:1987], [@toner:1998]), self-driven particles [@vicsek:1995], large collections of robots ([@reif:1999], [@spears:1999]), and multi-robot teams [@balch:1998].
 
 ## Related Software Packages
 
[…]
