+++ paper/paper.bib
@@ -180,6 +180,16 @@
   doi={10.3389/frobt.2017.00012}
 }
 
+@inproceedings{erez:2015,
+  title={Simulation tools for model-based robotics: Comparison of bullet, havok, mujoco, ode and physx},
+  author={Erez, Tom and Tassa, Yuval and Todorov, Emanuel},
+  booktitle={2015 IEEE international conference on robotics and automation (ICRA)},
+  pages={4397--4404},
+  year={2015},
+  organization={IEEE},
+  doi={10.1109/ICRA.2015.7139807}
+}
+
 @article{zoss:2018,
   title={Distributed system of autonomous buoys for scalable deployment and monitoring of large waterbodies},
   author={Zoss, BM and Mateo, D and Kuan, YK and Tokić, G and Chamanbaz, M and Goh, L and et al.},
