### 2023-07-28 mrsonandrade/pyswarming#4 @mrsonandrade
Excellent! Thank you for the feedback!

These are the points that I have worked on:
1. differentiating between the types of software related to swarm robotics;
2. organize the software between simulators and packages for behaviors;
3. mention what need PySwarming meets;
4. rolling in more discussion in the statement of need;
5. add something like "because of ABC, we introduce PySwarming, which does XYZ";
6. instead of "mid and final results", say "intermediate and final results";
7. add a repulsion component in the example;
8. adding more details about how a user might use this tool to create their own new behaviors;
9. talk about the "PySwarming is flexible and easy to extend" a little bit.

Let me know what you think.
And again 😃 , thank you for your time!!

PS: If you have an issue downloading the updated paper, you can access the [actions](https://github.com/mrsonandrade/pyswarming/actions/runs/5686829594) and download the pdf from the _Artifacts_ section.
