### 2023-07-28 mrsonandrade/pyswarming#4 @sea-bass
Thank you! I really like your latest changes, and especially good to see you added the repulsion behavior to your example.

Only 2 remaining comments:

* The table of related software packages is a nice addition; however, the paragraph is very long and difficult to follow. I suggest breaking it up a bit, though the content is fine. I also think your post-table paragraphs are better at the beginning, as they introduce the two types of packages before getting into the list. So I'd rearrange a little as well.
* Typo: "In addition, the PySwarming package comes with a simple read-to-use simulation feature" -- there is a typo, should be "ready-to-use".
