### 2023-07-28 mrsonandrade/pyswarming#4 @mrsonandrade
Great! I have split this paragraph into two: (1) from lines 31 to 39; and (2) from lines 39 to 55.

I did not get the idea in "your post-table paragraphs are better at the beginning". As I understood, are you talking about the paragraphs after the table (the statement of need)?
