### 2023-07-29 mrsonandrade/pyswarming#4 @sea-bass
> I did not get the idea in "your post-table paragraphs are better at the beginning". As I understood, are you talking about the paragraphs after the table (the statement of need)?

Yes, exactly! I think introducing the difference between simulation and behavior packages before going through the list and table may work better. Very minor.
