### 2023-07-30 mrsonandrade/pyswarming#7 @JHartzer
Auto-Differentiation Module
The module `auto_differentiation.py` seems to add significant complexity to the
package without much benefit since it is only used for the gradient calculation.
At the moment, it doesn't seem clear that to use the `geofencing` function with
more complex functions, those functions must be imported from
`auto_differentiation` e.g.
`from pyswarming.auto_differentiation import sin, cos, erf`
instead of using the existing math or numpy functions. A user (such as myself)
might attempt the following and be confused by the result
```python
>>> import numpy as np
>>> f = lambda x,y,z: np.sqrt(x+y+z)
>>> geofencing( (1,2,3), f)
array([nan, nan, nan])
```
Would it make more sense to utilize numeric gradient calculation to improve the
clarity of what is being done and allow users to create even more complex
functions using other libraries? I think that would produce results with all the
accuracy needed for the behaviors without needlessly increasing complexity and
limiting users in the construction of their geofencing functions.
