### 2023-07-30 mrsonandrade/pyswarming#8 @JHartzer
 [JOSS Review] Additional Paper Feedback
I think, like mentioned in the other review, there needs to be additional work
on the statement of need. In its current form, I don't believe the need is clear
or it aligns with the solution that is being presented. The cited paper by
Nedjah and Silva doesn't appear to call for this kind of work, but rather a
standard controller software that would be used on-machine with a standard
hardware platform. Since this work focuses on the congregation of swarming
behaviors that can then be combined into swarm algorithms, I don't think it
solves the need cited in this paper. However, I do think there is importance in
this work. Specifically calling out the need to quickly prototype and develop
new algorithms in python and leveraging existing datasets would more closely
align with the solution that is being proposed. This is mentioned later in the
paper as well, but I think it would strengthen the submission and this section
to add a similar note here.

I think the conclusion could be expanded as well. Both sentences are a bit
clunky, start with adverbs, and run on a bit. It might flow better to have a
more academic paper conclusion "presented in this work is pyswarming, a
package for... This package provides behaviors for swarming robotics...
Written in python, it is more flexible and allows access to datasets...
Therefore... etc."

I think improving these two sections would greatly improve the quality of the
paper, as the other sections are quite strong.

## Nitpicks:
- 'methos' should be 'methods'.
- 'ajustable' should be 'adjustable'
