### 2023-07-30 #5647 @JHartzer
In addition to the above comments, I have submitted the following blocking issue:
https://github.com/mrsonandrade/pyswarming/issues/8

And non-blocking issue:
https://github.com/mrsonandrade/pyswarming/issues/7
