### 2023-08-04 #5647 @adi3
@mrsonandrade @JHartzer @sea-bass after taking a deeper look into the codebase and discussing this with other editors, we have concluded that this paper _does_ meet the requirements of substantial scholarly effort. In examinig this, we also took setup, tests and examples into consideration as those are crucial parts of using OSS.

@mrsonandrade please inform @JHartzer when you're done fixing the remaining items so he can complete his review smoothly.
