+++ paper/paper.md
@@ -21,7 +21,7 @@ affiliations:
    index: 1
  - name: Ocean Engineering Program, Laboratory of Waves and Current, LOC/COPPE/UFRJ, Rio de Janeiro, Brazil
    index: 2
-date: 21 July 2023
+date: 04 August 2023
 bibliography: paper.bib
 
 ---
@@ -30,7 +30,7 @@ bibliography: paper.bib
 # Summary
 
 When considering a system composed of a group of robots, swarm robotics is an approach that can be used to coordinate this group. These swarms can be inspired or not by social insects or other animal societies [@trianni:2008], where basic behaviors are usually used to compose complex tasks. These previously mentioned basic behaviors have been studied for a long time, with applications, for example, to flocks, herds, and schools [@reynolds:1987] and multi-robot teams [@balch:1998].
-However, software standardization is urgently needed to maximize the applicability of swarm robotics to real-world applications [@nedjah:2019]. Here we introduce PySwarming, a tool that makes easy the coordination of swarms and serves as a centerpiece, organizing different methods developed in the swarm robotics field. Its flexibility (written in Python) and customizability (easily customized by users) encourage interaction and scientific progress in the research community.
+Here we introduce PySwarming, a tool that makes easy the coordination of swarms and serves as a centerpiece, organizing different methods developed in the swarm robotics field. Its flexibility (written in Python) and customizability (easily customized by users) encourage interaction and scientific progress in the research community.
 
 # Introduction
 
@@ -39,7 +39,9 @@ Additionally, these swarms may or may not be inspired by social insects and othe
 
 ## Related Software Packages
 
-For more than 20 years software for swarm robotics has been created, adapted, and tested in different ways [@calderon:2022].  Software packages that allow the creation of virtual scenarios and swarm robots with sensing, processing, and actuating capabilities are crucial for the swarm robotics field. These software packages may be split into two categories: (1) *Swarm Simulators*, which in general can simulate swarm robots with sensors and actuators, and (2) *Behavior Packages*, which are composed of a bunch of ready-to-use collective behaviors. Also, these software are written in different programming languages, which may be an initial barrier for new users, depending on how much this language is difficult to learn or previous knowledge by the user. Concerning swarm robotics, [@calderon:2022] presents a good review regarding swarm robotics simulators, platforms, and applications. Therefore, for software packages we have, for instance: (1) Buzz, which is a programming language for heterogeneous robot swarms [@pinciroli:2015]. It offers primitives to define swarm behaviors and also single-robot instructions; (2) ChoiRbot by [@testa:2021], which is a toolbox for distributed cooperative robotics based on the Robot Operating System (ROS) 2; (3) ROS2Swarm, which is a package for applications of swarm robotics that provides a library of ready-to-use swarm behavioral primitives [@kaiser:2022]; (4) ARGoS, which is a multi-physics robot simulator, able to simulating large-scale swarms of robots [@pinciroli:2011]; (5) Stage Simulator, which provides a virtual world populated by mobile robots and sensors, along with various objects for the robots to sense and manipulate [@vaughan:2008]; (6) USARSim is a free 3D simulator similar to Gazebo [@carpin:2007]; (7) The Swarm-bots project [@mondada:2004] is also a robotic simulator with swarm-intelligence-based control mechanisms, but it is not publicly available; (8) and TeamBots is a Java-based collection of application programs and Java packages for multiagent mobile robotics research [@balch:1998teambots]. **Table 1** shows a summary of these software and their respective category and programming languages.
[…]
