+++ paper/paper.bib
@@ -222,14 +222,6 @@
   doi={10.1023/B:AURO.0000033970.96785.f2}
 }
 
-@article{nedjah:2019,
-  title={Review of methodologies and tasks in swarm robotics towards standardization},
-  author={Nedjah, N and Junior, LS},
-  journal={Swarm and Evolutionary Computation},
-  year={2019},
-  doi={10.1016/j.swevo.2019.100565}
-}
-
 @book{trianni:2008,
   title={Evolutionary Swarm Robotics},
   author={Trianni, V},
