### 2023-08-04 mrsonandrade/pyswarming#8 @mrsonandrade
Phenomenal comments @JHartzer ! Thank you!
By following your points, I have worked on the statement of need and the conclusion. Let me know what you think 😃
