### 2023-08-10 mrsonandrade/pyswarming#7 @mrsonandrade
@JHartzer thank you for the observation! I looked here and I have found the `numdifftools` [package](https://github.com/pbrod/numdifftools). I will make some tests and if it works as expected I will migrate to it 😃 As you said, I think that this will improve the clarity of what is being done.
