### 2023-08-18 #5647 @mrsonandrade
Thank you @adi3!

Just to inform all of you that two papers related to PySwarming were accepted for publication in the 20th IEEE Latin American Robotics Symposium. As soon as possible, I will insert their information on the README file from PySwarming.

I have worked on the blocking issue from @JHartzer.

Regarding the non-blocking issues from @JHartzer and @sea-bass, I am planning to work on them as soon as possible. I have a new contributor in my lab that will work with this package, so, I hope this will help to work on the non-blocking issues and improve the package.
