### 2023-09-26 ae33424
Fixes multi-reference citation in JOSS paper
