### 2023-09-26 #5647 @kyleniemeyer
Hi @mrsonandrade, I'm doing some final checks before publishing.

I made a few fixes to references in this PR, could you merge it? https://github.com/mrsonandrade/pyswarming/pull/9
